move segment normal to utils
