+++ examples/add_vectors_image.py
@@ -15,7 +15,7 @@ with app_context():
     # create the viewer and window
     viewer = ViewerApp()
 
-    image = 0.2*np.random.rand(24, 100) + 0.5
+    image = 0.2*np.random.random((24, 100)) + 0.5
     layer = viewer.add_image(image, clim_range=[0, 1], name='background')
     layer.colormap = 'gray'
