+++ napari/layers/_vectors_layer/model.py
@@ -7,6 +7,7 @@ from .._base_layer import Layer
 from .._register import add_to_viewer
 from ..._vispy.scene.visuals import Mesh
 from ...util.event import Event
+from ...util import segment_normal
 from vispy.color import get_color_names
 
 from .view import QtVectorsLayer
@@ -434,29 +435,3 @@ class Vectors(Layer):
 
         self._need_visual_update = True
         self._update()
-
-
-def segment_normal(a, b):
-    """Determines the unit normal of the vector from a to b.
-
-    Parameters
-    ----------
-    a : np.ndarray
-        Length 2 array of first point
-    b : np.ndarray
-        Length 2 array of second point
-
-    Returns
-    -------
-    unit_norm : np.ndarray
-        Length the unit normal of the vector from a to b. If a == b,
-        then return [1, 0]
-    """
-    d = b-a
-    normal = np.array([d[1], -d[0]])
-    norm = np.linalg.norm(normal)
-    if norm == 0:
-        unit_norm = np.array([0, 0])
-    else:
-        unit_norm = normal/norm
-    return unit_norm
