+++ napari/util/misc.py
@@ -141,3 +141,29 @@ def formatdoc(obj):
         return obj
     finally:
         del frame
+
+
+def segment_normal(a, b):
+    """Determines the unit normal of the vector from a to b.
+
+    Parameters
+    ----------
+    a : np.ndarray
+        Length 2 array of first point
+    b : np.ndarray
+        Length 2 array of second point
+
+    Returns
+    -------
+    unit_norm : np.ndarray
+        Length the unit normal of the vector from a to b. If a == b,
+        then returns [0, 0]
+    """
+    d = b-a
+    normal = np.array([d[1], -d[0]])
+    norm = np.linalg.norm(normal)
+    if norm == 0:
+        unit_norm = np.array([0, 0])
+    else:
+        unit_norm = normal/norm
+    return unit_norm
